+++ src/lib/components/HomePage/HomePage.test.tsx
@@ -1,0 +1,17 @@
+import React from "react";
+import renderer from 'react-test-renderer';
+import { HomePage } from "./HomePage.jsx";
+
+it("renders the HomePage component with the children text", () => {
+    const component = renderer.create(
+        <HomePage
+            appLogo={""}
+        >
+            <p>HomePage Children text 123</p>
+            <p>HomePage Children text 456</p>
+            <p>HomePage Children text 789</p>
+        </HomePage>
+    );
+    let tree = component.toJSON();
+    expect(tree).toMatchSnapshot();
+});